requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
tenacity==8.2.3
psutil==5.9.6
python-magic==0.4.27
google-cloud-storage>=2.12.0
mysql-connector-python>=8.3.0
